Code cleanup utilize rubocop for style and consistency

+++ tools/clean.sh
@@ -26,8 +26,7 @@ shift $((OPTIND-1))
 
 # Very simple user error checking
 if [ "$user" = '' ]; then
-  echo "A username required... bailing."
-  exit 1;
+  echo "No user specified, I'm just going to remove the apps."
 fi
 
 if [ "$(whoami)" != 'root' ]; then
@@ -36,7 +35,7 @@ if [ "$(whoami)" != 'root' ]; then
 fi
 
 echo 
-echo "I am about to remove user account $user and its home directory."
+echo "I am about to remove Puppet-agent, VBox, and Vagrant, and a user account (if specified)."
 echo -n "Are you sure you want to do this? (YES/no), then enter: "
 read answer
 
@@ -54,12 +53,13 @@ pkgutil --forget com.puppetlabs.hiera
 for f in $(pkgutil --only-files --files com.puppetlabs.facter); do rm /$f; done
 for d in $(pkgutil --only-dirs --files com.puppetlabs.facter | tail -r); do rmdir /$d; done
 pkgutil --forget com.puppetlabs.facter
-for f in $(pkgutil --only-files --files com.puppetlabs.puppet-agent); do rm /$f; done
-for d in $(pkgutil --only-dirs --files com.puppetlabs.puppet-agent | tail -r); do rmdir /$d; done
+rm -rf /opt/puppetlabs
 pkgutil --forget com.puppetlabs.puppet-agent
 
 $DIR/vagrant-uninstall.tool
 $DIR/vbox-uninstall.tool
 
-dscl . -delete /Users/$user
-rm -rf /Users/$user
+if [ "$user" != '' ]; then
+  dscl . -delete /Users/$user
+  rm -rf /Users/$user
+fi
